requirements.txt
PyQt5>=5.15.0
pyinstaller>=5.0.0
Pillow>=8.0.0
